requirements.txt
flask
openai
load_dotenv
flask-cors
requests